Buidling a fast but persistence timed set
Idea is to use redis for persistence and add the elements to it in a non blocking way. But using the Set for operations.
In this form it is 100x faster than RedisSet. But it is still 10 times slower than Set.
Slowness compared to Set must be related to creating a gorouting for every insert.